clean up modules/ fix newGame func/ fix import-exp
