+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Tic Tac Toe</title>
-    <script type="module" src="./js/ui-module.js" defer></script>
+    <script type="module" src="./js/main.js" defer></script>
     <link rel="stylesheet" href="css/style.css" />
   </head>
   <body>
